phonebuzz3/requirements.txt
twilio
django-twilio
django-phonenumber-field
celery
django-durationfield